http://bugzilla.zimbra.com/show_bug.cgi?id=16422 Bug: 16422
Turn off "double click isolation" for hybrid view with
reading pane in Firefox. We instituted a delay after a
single click so that if there turned out to be a double
click (in order to load conv/msg into full view), we wouldn't
waste time loading the msg into the reading pane (which
would flash).

The annoying side effect is that there's a 300ms delay on
a regular single click. Since we now have a shortcut ("o") to
open a conv/msg into full view, we don't need to optimize
dbl-click so much.