http://bugzilla.zimbra.com/show_bug.cgi?id=16476 http://bugzilla.zimbra.com/show_bug.cgi?id=16478 bug: 16476, 16478
Allow incremental backup to succeed when archived redo logs have gaps in sequence relative to the logs already backed up.  This is considered a partial success, and the backup report email will mention the problem.

Added zmrestore --ignoreRedoErrors option (also in RestoreRequest SOAP command).  This option is off by default.  If specified, any and all errors during redo log replay portion of restore will be logged and ignored.  This option became necessary because we now allow missing redo logs in backups.  If logs are missing and this option is used to restore, the restored mailbox won't have all expected data.  Use this option only in emergency when mailbox consistency is not important and your primary concern is data recovery.