http://bugzilla.zimbra.com/show_bug.cgi?id=16789 Bug: 16789
Kbnav: check init during enable()

When I looked at this problem via Firebug on a Mac that
had lost kbnav, the keyboard mgr appeared not to have
been initialized. I'm not sure how that could happen. Anything
that uses kbnav has to enable it, so I added a check for init
status there.