bug: 824
Add server support for high- and low-priority messages.  They're stored as (hopefully) mutually-exclusive FLAGS bits and are returned in the "f" attribute of the <m> as "!" and "?" respectively.  You can also set the "f" attribute to "!" or "?" on the <m> of the SendMsg request to set the outbound message's priority.

We look at X-Priority first, then Importance.  Outbound messages include both.  "!" has priority over "?" when sending a message.

http://bugzilla.zimbra.com/show_bug.cgi?id=824