bug: 15724
In metadata of each Invite object record if the owning account was the organizer account at the time of invite creation.  This way organizer check no longer depends on email address at check time, and allows preserving organizer status even after account email address is changed.  This new attribute is set only on newly created or edited invites.  It is tracked at invite level rather than appointment level because technically it is possible for a single appointment to have different exception invites with different organizers as per RFC2446.

This fix will help with modifying personal appointments.  For group appointments, this "fix" creates a new "bug" case because when the renamed account modifies an appointment and sends out update invite, the new invite will show the new email address as organizer, and the attendee mailbox will see this as a spoofing attempt and refuse to process the invite.  There is no good way to solve this problem because the attendee has no way to know that the old and new organizer addresses are different as a result of a legitimate rename operation.  This issue is related to bug 7129.  For the present bug, addressing the personal appointment case is good enough.

http://bugzilla.zimbra.com/show_bug.cgi?id=15724