Bug: 12212, 13764
Two sets of changes which touch a lot of the client:
1. Handle accounts (family mailbox)
2. Rewrite overview logic


* We now have ZmAccount objects, which were necessary to support family mailbox. In the
default case where family mailbox is not enabled, there is a single account which
represents the authenticated user. The account holds its own copy of a user's settings
and data trees (folders, tags, and zimlets). Accessor methods in ZmAppCtxt have been
updated to be aware of the account. They default to the current active account.

* One important aspect of child accounts is that the folder IDs are composite. They
consist of the account ID and the local numeric folder ID joined by a colon. The biggest
impact of that comes when we try to figure out if a folder is a special (system) folder
by its ID. To help with that, folder objects have a new property "nId", which you can
think of as "numeric ID" or "normalized ID" that is just the local part, so it can be
compared to known system folder IDs.

* A new class, ZmOverview, has been created. It contains a lot of view-related
code that ZmOverviewController used to have. An overview is a set of tree views.
An overview maintains its own options (which are applied to the tree views). Those
options used to be in a bunch of hashes in ZmOverviewController.

Tree views are no longer pruned from and readded to the DOM. We used to do that as
an optimization, and maintain a single overview for all the apps by shuffling tree
views in and out. Now that each app has its own overview, that's no longer necessary.

* Apps manage their own overview content. That content could be a ZmOverview, a
DwtAccordion, or a custom DwtComposite. The app is free to create what it wants
displayed in the overview pane. By default, an overview with the app's registered
list of tree views is created. ZmApp had added the following methods:

getOverviewPanelContent()

	Returns a DwtControl with the app's overview content

setOverviewContent()

	Replaces content in the overview pane with this app's overview content,
	using the new method ZmAppViewMgr::setComponent()

getOverviewPanelContentId()

	Defaults to app name

getOverview()

	Returns the app's current ZmOverview, if any

getOverviewId()

	Returns the ID of the app's current ZmOverview, if any

_getOverviewParams()

	Returns a hash of default overview options

_getOverviewTrees()

	Returns the list of tree views (organizer types) to show

_getHideEmpty()

	Returns a hash of which trees to hide when they have no data

_accordionSelectionListener()

	Does some basic checking. App should override to take action.

_activateAccordionItem(item)

	Expands the given accordion item, creating its overview if needed


* Organizer types FOLDER and SEARCH are now registered by ZmMailApp, and TAG is
registered by ZmZimbraMail.

* There's now an overview option to not show tooltips (now that we have them).

* Family mailbox: switching accounts runs initial search. At some point we may want
to do some caching/view preservation. Also, the advanced search is reset, since there
may be pickers showing folders from the previous account. User info in the upper right
gets reset to the new account.


- ZmZimbraMail._OVERVIEW_ID is gone

- Added DwtAccordion::setItemContent, which attaches a control's HTML element to
an accordion item

- Renamed DwtAccordianItem to DwtAccordionItem :p

- Added ZmAppCtxt::getCurrentApp()

- Factored isShared() up to ZmItem

- Removed ZmZimbraMail::_checkOverviewLayout() and ZmZimbraMail::_getOverviewTrees()

http://bugzilla.zimbra.com/show_bug.cgi?id=12212
http://bugzilla.zimbra.com/show_bug.cgi?id=13764