Bug: 14156
Lots of cleanups and simplifications came along with the new overview
design (and there's still more to come).

ZmRosterTreeController is now a plain object that contains some
handlers.  It'll probably merge with ZmChatListController at some
point.

Got rid of ZmRosterTree, ZmRosterTreeItem and ZmRosterTreeGroup.

Added view/ZmImOverview.js -- a DwtComposite which contains all that IM
app needs to display in the sidebar.  This object hooks to the main
ZmRosterItemList events, in order to keep the buddy list view updated.
It is possible to instantiate more such objects, should this ever be
needed, and they will keep synchronized with what's in the server.

ZmImOverview also creates the buddy list action menus and calls
the appropriate listeners in ZmRosterTreeController (for now).

TODO: keep the list ordered, don't load IM from the start.

http://bugzilla.zimbra.com/show_bug.cgi?id=14156