bug: 15260
Fixed <replies> generation in GetAppointmentResponse to return it as child of <appt> element, rather than child of <inv> or <comp>.  Replies need to be at the top level because they don't necessarily have a one-to-one mapping with invites.

SetAppointmentRequest can now specify <replies>, as sibling of <default>, <except>'s and <cancel>'s.  If no <replies> element is given, existing replies are kept.  If <replies> element is present, appointment's reply list is replaced with the specified replies.  Special case: <replies/> with no child <reply> elements means to clear the existing replies from the appointment.

This required changes to SetCalendarItem redo op, so redo log version went up to 1.15.

http://bugzilla.zimbra.com/show_bug.cgi?id=15260