bug: none
Simple perl script that makes it easy to parse out a small number of columns from mailboxd.csv (previously zimbrastats.csv...why do we keep changing the name and location of everything anyway??).  

The script scans the csv file and finds the 'header row' to match to the requested column: so it won't get out of date when we add/remove stats.  Running the script with just a filename will print all available counters.


Example:

% csvCols.pl mailboxd.csv -5 timestamp mpool_perm_gen_free heap_free
          timestamp, mpool_perm_gen_free, heap_free, 
06/07/2007 19:19:38,              235856,  16557304, 
06/07/2007 19:20:38,              235856,  16347640, 
06/07/2007 19:21:38,              235856,  16149128, 
06/07/2007 19:22:38,              235856,  15948584, 
06/07/2007 19:23:38,              235856,  15750120,