bug: 17133
Added includeContent="1" option to GetAppointmentRequest (and GetTaskRequest) to fetch body parts containing meeting notes inline.  An <mp> will be inserted for the top level MIME part, as a child element of <inv> and sibling of <comp>.  More <mp>'s can be nested to match the MIME parts.  text/calendar parts are excluded because they're already represented by the <inv>.  Notice the location of <mp> as child of <inv> is different from its location in a GetMsgResponse, where it's a child of <m> and sibling of <inv>.  This discrepancy is necessary because GetAppointmentResponse does not use <m> element.

As with GetMsgResponse, attachments only return skeletal <mp> elements without the content bytes.  They have to be fetched in separate requests.

http://bugzilla.zimbra.com/show_bug.cgi?id=17133