bug: 15833
Added ability to sort task search results by Due-Date, Status, and Completion-%age

For <SearchRequest> with types="task" (and *only* task), the following additional sorts are valid:
                    taskDueAsc|taskDueDesc|taskStatusAsc|taskStatusDesc|taskPercCompletedAsc|taskPercCompletedDesc

Note that the server implements task-sorting by buffering all the hits in memory (up to a limit of 5,000) and then re-sorting them in memory.  As such, if the user has more than 5000 tasks returned from the specified search, the results are undetermined.

http://bugzilla.zimbra.com/show_bug.cgi?id=15833