bug: 16633
Allow the client to specify ci="{content-id}" on an <mp> in SendMsg.

Allow the client to insert an attachment in a particular location in SendMsg by inlining an <attach> element in the body <mp> instead of specifying the <content> directly.  (Note that in this case, the ct attribute on the part is ignored in favor of the Content-Type associated with the attachment.)

http://bugzilla.zimbra.com/show_bug.cgi?id=16633