Bug: 17734
New login page for Ajax client. To use it, go to

    /zimbra/public/unifiedLogin.jsp

It will forward to a new JSP: launchZCS.jsp

The old login page is still available, and is still set as
the welcome page:

    /zimbra/public/Login.jsp

It forwards through some JS code and a couple of
servlets to launchZimbraMail.jsp. ZmZimbraMail is set up
to handle launch from either login path.

http://bugzilla.zimbra.com/show_bug.cgi?id=17734