bug: 17772
So that we use log4j for jetty logging.  That way everything is controlled by log4j.properties.

Had to patch ical4j-0.9.16-patched.jar because the version we have somehow has a log4j.properties in the jar that masked the log4j.properties we want to load.

Also to avoid dependency on ZimbraPatternLayout we use a separate log4j.properties on jetty classpath first, until SoapServlet resets log4j.properties.

http://bugzilla.zimbra.com/show_bug.cgi?id=17772