bug: none
JJ's change P4 51699 changed the server so that it used Jetty's slf4j.jar, instead of the one we compiled against -- unfortunately Jetty is using a different version of sl4j than what we have checked into the tree.  The two were not binary-compatible and so this was causing IM to break in a bunch of ways:  couldn't connect to Interop services, couldn't connect via 3rd party clients.