bug: 16607
Renamed zimbraBackupAutoGroupedNumIntervals global config attribute to zimbraBackupAutoGroupedNumGroups.

Backup interval can be specified in weeks or months in addition to days and hours/minutes/seconds.  Each of days, weeks and months must be specified by itself and cannot be combined with anything else.  Hours, minutes and seconds can be combined for easier testing.

http://bugzilla.zimbra.com/show_bug.cgi?id=16607