bug: 17758
Support delta sync constrained to a single folder and its subfolders by specifying l="{sync-root}" on the delta SyncRequest.  This can be done on the authenticated user's mailbox or another user's mailbox.  Items that fall outside the sync root are serialized as *deletes*.

Fix a bug that was causing us to fetch and serialize way too many items in large delta sync responses.  We were building up batches of items to sync, then operating on the whole modified list each time.

http://bugzilla.zimbra.com/show_bug.cgi?id=17758