bug: 16607, 17192
Big change in how deleted accounts are handled by backup/restore.  Previously, every full or incremental backup was noticing deleted accounts/mailboxes by comparing the list of accounts requested for backup and the list of accounts in LDAP.  Accounts that are missing from LDAP must be deleted ones, and these are recorded in the backup.  The problem with this method is that deleted accounts are noticed only upon backup.  During restore, it is not possible to detect a deleted account if no backup was done after account was deleted.  This is the issue in bug 17192.

A bigger problem arises with auto-grouped full backups.  In an auto-grouped backup, account list in LDAP aren't examined and mailboxes to be backed up are selected out of those that are currently in the zimbra.mailbox table, so it's impossible to determine which mailboxes have been deleted since the last backup.

The solution is to add deleted_account table in the database.  Every time an account/mailbox is deleted, we insert a row in this table.  If the mailbox is restored (and thus gets reinserted to mailbox table), we delete the row from the deleted_account table.  The primary key in this table is the email address, because email is how accounts are identified in backup/restore commands.

Database version is incremented to 44.  Backup version is incremented to 6.2, invalidating old backups.

http://bugzilla.zimbra.com/show_bug.cgi?id=16607
http://bugzilla.zimbra.com/show_bug.cgi?id=17192