bug: 18302
The client needs the task's "Due Date" returned in search results.  For now, add a 'dueDate' attribute to the <task> element with the correct info.  

It might be possible to use the MailItem's Date field for storing the task's due date.  This would make DueDate sorts faster and allow Task searches to use the DATE: operator when searching.

http://bugzilla.zimbra.com/show_bug.cgi?id=18302