bug: 6937, 17767
Full support for mountpoints in IMAP.  You can't create mountpoints via IMAP, but they're fully functional from IMAP once created via another interface.  The only exception is that you can't SELECT/EXAMINE a mountpoint that points to a folder hosted on another server.

Other people's mountpoints are hidden, as are mountpoints that point back to your own folder -- this is to make things simpler and remove the possibility of infinite loops in LIST processing.  We also hide mountpoints for which you don't have read access on the mountpoint's target.

The key concept in this implementation is that an ImapPath has a "referent", the thing it points to.  A normal ImapPath has its referent point to itself.  A mountpoint has its remote target folder as its referent.

http://bugzilla.zimbra.com/show_bug.cgi?id=6937
http://bugzilla.zimbra.com/show_bug.cgi?id=17767