bug: 13827
Allow CreateContact, ModifyContact, and ContactAction op="update" to specify contact attachments already in the store.  We support attaching messages (or their subparts), contacts (or their attachments), and documents.  The syntax for attaching an existing store item to a contact is

    <a n="{attr-name}" id="{item-id}" [part="{subpart-name}"]/>*

Note that we default "id" to the contact being edited in ModifyContact and ContactAction.

http://bugzilla.zimbra.com/show_bug.cgi?id=13827