Bug: 17027
Fix inconsistency with which msg is displayed when a conv
is shown. Previously, conv view would show the first
matching msg, and conv list view would show the first msg.
Now they both show the first matching msg.

When a conv is selected in CLV (via arrow or mouse click),
we now do a SearchConvRequest instead of a GetMsgRequest.
Since we do a SearchConvRequest when a conv is expanded,
we have probably reduced our number of SOAP requests for
standard use of CLV.

http://bugzilla.zimbra.com/show_bug.cgi?id=17027