bug: 16942
Official support for document versioning.  Each item has a 1-based version number that's maintained in the METADATA column and is incremented each time it's renamed or its content is changed.  (Still may be a few more places to patch this...)  Old revisions' data are stored in the new REVISION table.  Introduced the idea of an item's "total size" (item plus all saved revisions), which is calculated rather than persisted and is used to calculate folder and mailbox sizes.  The versioning support is at the MailItem level, so theoretically any item could be versioned using the REVISION table.  Also added a new persisted "versioned" flag as well as a non-persisted "uncached" flag used to force the server not to cache non-current revisions.

At present, saved revision limits for documents and wiki items are controlled by the zimbraNotebookMaxRevisions LDAP attribute; when set to 0, all old revisions are maintained.  If configured to allow a maximum of 2 revisions for contacts, the server could theoretically provide for undo on contact edit.

The database schema version is now 45.  Please either reset-the-world or run the migrate20070725-CreateRevisionTable.pl migration script.

(Not sure if this is intentional or not, but when a caller requests a nonexistent version of a wiki page from the WikiFormatter, the current version is returned.)

Does not support backup and restore of item revisions.

http://bugzilla.zimbra.com/show_bug.cgi?id=16942