bug: 14812
Allow the caller to constrain the date returned in the GetInfo response by specifying the "sections" attribute on the GetInfoRequest.  This attribute is a comma-separated string list; possible entries in this list are "mbox", "prefs", "attrs", "zimlets", "props", "idents", "sigs", "dsrcs", and "children".  Omitting a section name from this list causes the corresponding part of the GetInfo response to be omitted.  By default, all sections are returned.

http://bugzilla.zimbra.com/show_bug.cgi?id=14812