bug: 14535
Major cleaning up of how we build zdesktop.  Now it's possible to make a dev build to install under /opt/zimbra/zdesektop so that it can co-exist with a regular dev build under /opt/zimbra.  Also make sure we share as much build script as we can between the dev build and the installer build.  Now there's no material difference between a dev install and an installer install.

http://bugzilla.zimbra.com/show_bug.cgi?id=14535