bug: 18981
do not reload in checkAccountStatus.  

Staled data can be read back if there are replication delays and the account is refreshed from a not-caught-up replica.  Since all mods (via SOAP) are proxied to the home server and we refresh from the master after each mod,  the cached entry should contain the most updated data.

http://bugzilla.zimbra.com/show_bug.cgi?id=18981