bug: 7074
The server now detects if a single client AuthToken is attempting to make two simultaneous instant_notify requests at once.  If that happens,  the server cancels the older request with a flag to the client 'waitDisallowed'.

Updated the client to look for the 'waitDisallowed' flag -- if it receives the flag, the client stops trying instant-notify and goes back to polling.

This should clear up instances of the browser hanging w/ instant_notify (b/c of the two-connections-per-server limit in the browser) if you tried to open up a second browser window to the server -- instant notify should be disabled for all but the most recent app instance.

http://bugzilla.zimbra.com/show_bug.cgi?id=7074