bug: none
On dogfood today, an uncaught  NPE in ConversionServlet's cleanup timer caused the shared-across-the-system Timer thread to die.  The loss of the timer thread caused session cleanup to stop,  servlet cleanups to stop, waitsets cleanup to stop...and a whole host of other things -- eventually leading to an OOME on Dogfood.


Two fixes:

- Fix the actual code in the ConversionServlet that threw the NPE.  (null check)

- In all the places we use the Zimbra.sTimer shared timer, wrap the timer execute code in a try{}catch block and catch (and log) any exceptions....we really don't want the timer thread dying just because a timer client threw.


A better solution would probably be to wrap java.util.Timer with something that did a try/catch around TimerTask.run....couldn't quite figure out the right way to do it though, so leaving it as  TBD.