bug: 18889, 15636
Since we're not using the conversation's "subject" to serialize a conversation search result hit, there's no longer any reason to track the first message in the conversation and to recalc the metadata when that may have changed.  The converation now only tracks its own normalized subject (in the SUBJECT database column).

When a conversation is loaded and the METADATA has been nulled out because of an empty trash, we no longer reload all the messages to constitute the sender list.  That is now done only when requested.

Also, when a message is removed from the conversation and we don't have a SOAP session active, we just write NULL to the METADATA column rather than recalculating the sender list.  Similarly, if the Trash is being emptied (or old messages are being purged) and there's no active SOAP session, don't bother loading all the modified-but-not-removed conversations because nobody cares at the moment.

http://bugzilla.zimbra.com/show_bug.cgi?id=18889
http://bugzilla.zimbra.com/show_bug.cgi?id=15636