bug: 19884
The convertd code had logic to drain and close the PipedInputStream that sometimes gets generated by the activation framework when you ask JavaMail for a MimePart's InputStream.  Move this logic to ByteUtil.closeStream() and make sure that all places that get an InputStream from a DataSource call it in a finally block.

http://bugzilla.zimbra.com/show_bug.cgi?id=19884