Bug: 17090, 19970
Split loading of main packages to before and after inbox
has been displayed.

Before this checkin, we had split startup files into two packages,
Startup1 and Startup2, but were loading them both up-front
to be safe. Now, Startup2 is not loaded until the mail app has
rendered.

* ZmZimbraMail maintains a list of "post-render" callbacks to run
after the start app (mail) has rendered. There is a POSTRENDER
app event to handle that. During startup, callbacks can be added
to a list to run after rendering has occurred. There's a listener for
the POSTRENDER event that does that as a timed action (so that
the UI will have been painted). The current list of post-render
callbacks:
    - load Startup2 package
    - load Zimlet package, init zimlets
    - run app startup functions

* On my localhost, delaying the load of Startup2 saved around
2 seconds. Note that there is no network cost.

* Startup1 has been split into two roughly equal packages, as FF
especially seems to take exponentially long to parse large piles of JS.
The original parse time of around 1800ms got cut in half.

* Move some constants from ZmCalItem to ZmCalendarApp so
that they're available to other apps if CalendarCore has not been
loaded.

http://bugzilla.zimbra.com/show_bug.cgi?id=17090
http://bugzilla.zimbra.com/show_bug.cgi?id=19970