bug: 17956
If any incoming ics object has SUMMARY value with "\n" in it, replace it with a space and make sure the value is a single line, as (loosely) stated in RFC2445.  Conversion will unescape the value and replace each sequence of CR and/or LF with a single space character.

http://bugzilla.zimbra.com/show_bug.cgi?id=17956