bug: 19234
Force analysis of ParsedMessage object before uploaded files are deleted.  If we don't force it the analysis happens on demand, and if called after deleting uploaded files, we can get IOException in JavaMail code as it tries to read from an InputStream attached to a deleted file.

The error was happening during appointment creation with attachments when notes was non-empty.  In this case ParsedMessage analysis was getting deferred to well after upload deletion.  But if notes was blank, the code tries to get "fragment" from ParsedMessage because it couldn't determine a suitable fragment from blank notes on the ics object, and this had the effect of triggering ParsedMessage analysis and JavaMail processing before upload deletion.

http://bugzilla.zimbra.com/show_bug.cgi?id=19234