bug: 18571
An admin doing a delegated mail send doesn't have to specify a Sender header, in which case none is used (the "send-as" case).  But if they *do* specify themselves in the Sender header (the "on-behalf-of" case), we no longer force it into the "send-as" case.

Likewise, when an admin does an "on-behalf-of" instead of a "send-as", we save a copy to the on-behalf-of user's mailbox as well.  We used to treat this the same as the send-as case, which suppresses save-to-sent for the sent-as user.

http://bugzilla.zimbra.com/show_bug.cgi?id=18571