Bug: 15964
Create a package for the JS needed by new window
(both compose and RFC822 display). The package is
split into two subpackages for faster loading.

A lot of JS is loaded, and further optimization could be
done.

Added a bunch of checks for child window to avoid loading
classes that it doesn't need. Since the package contains
selected mail classes, we pretend we've loaded MailCore and
Mail so they don't really get loaded (and bring in other
dependencies we don't want).

- Moved ZmAttachDialog from Share package to Startup2

http://bugzilla.zimbra.com/show_bug.cgi?id=15964