Bug: 17090
When we get a <refresh> block during startup, add its
handling to the list of things to do after Inbox has rendered.
Had to do a bit of shuffling in startup so that we know that
we're doing post-render callbacks.

The visual impact is that the overview will render after the
list view.

http://bugzilla.zimbra.com/show_bug.cgi?id=17090