bug: 18951
When sending an unsolicited BYE because we're dropping the connection due to inactivity (or some other reason), set a 10-second timer in another thread.  If the output stream hasn't been closed and cleared by the time the timer expires, that thread just closes the stream.  That should unblock the BYE and allow the main thread to proceed.

http://bugzilla.zimbra.com/show_bug.cgi?id=18951