bug: 20640
fix "is:remote" for cases where the shared folder has subfolders (is:remote now does the expected thing, and returns results in subfolders of mounted remote folders)

added "UNDER:" search operator -- it is just like "IN:" except it also includes matches in subfolders.

note that mutually refering folder hierarchies ARE possible, and they currently WILL bring the server down if you search recursively in them....so don't do that right now, until the bug is fixed.

http://bugzilla.zimbra.com/show_bug.cgi?id=20640