bug: 18620
Backing out the change in revision 120 because it made private appointments completely invisible in delegate user's calendar view.  Delegate should be allowed to there is something, i.e. the user is busy, without knowing what the appointment is about.  Unfortunately, this reintroduces bug 18620, that a private appointment is searchable by keywords.

http://bugzilla.zimbra.com/show_bug.cgi?id=18620