bug: 17712, 19928
When calendar resource auto declines a meeting request due to conflict, the reply email shows the conflicting meeting times.  These are shown using the invite's timezone.  But sometimes the invite will use UTC/GMT timezone.  When that happens, use the timezone stored in account preferences of the organizer, if the organizer is a local user.  If organizer is not a local user, use the resource's timezone.

http://bugzilla.zimbra.com/show_bug.cgi?id=17712
http://bugzilla.zimbra.com/show_bug.cgi?id=19928