bug: 20616
Observe user setting on sync interval for datasource sync.  Do the same as we do with zimbra sync on auth failure retry and offline retry.  For offline datasource we use a different attribute, offlineSyncInterval, for interval setting instead of the zimbraDataSourcePollingInterval attribute used for online version.  This is mainly to avoid triggering ScheduledTask stuff because we use our own sync model.

http://bugzilla.zimbra.com/show_bug.cgi?id=20616