bug: 17031 - POP3 data source: malformed date header causes SQLException on insert
Now checking the timestamp value before adding the message.  If it's
out of the valid range, we use the current date.

Enabled POP3 cleartext login for dev build.

Factored out data source import test code into TestUtil so that it
can be used by both the POP3 and IMAP tests.

http://bugzilla.zimbra.com/show_bug.cgi?id=17031