bug: 19235
Added DEBUG-ONLY support for batched-behind message indexing.  This code does not persist the list of unindexed message ids (it just stores the list in memory) and therefore should not be used in any kind of a production environment.  We can however use this to get some perf numbers and see if the batched index approach is helpful at all.

Batched indexing is enabled by setting the localconfig key "debug_batch_message_indexing"  to some number, suggest between 50-500 for testing purposes. 

http://bugzilla.zimbra.com/show_bug.cgi?id=19235