bug: 20616
Added local mailbox SMTP sender.  Similar to offline mailbox, we do a 2 stage send.  First UI SendMsg to local mailbox which dumps it into Outbox.  Then the sweeper tries to SMTP send the message.  Yet again javamail proves to be a piece of !@#$%.  Took forever to figure out SMTP over TSL with auth to Gmail SMTP but finally it worked.

http://bugzilla.zimbra.com/show_bug.cgi?id=20616