Bug: 14156
This will break your IM addresses in contacts.  Ha ha.

Until now, in order to add i.e. an Yahoo address to some contact, and
make it interoperate with the IM app, you had to enter:

  screen_name@yahoo.zimbra.com

Or, on your dev. machine:

  screen_name@yahoo.localhost.localdomain (or something)

Aside for being ugly and confusing, the biggest problem is that the
hostname of the Zimbra server needed to be a part of the address.

Now (only in contacts!) we store addresses like this:

  yahoo://screen_name
  local://foo@zimbra.com
  msn://msn_screen_name
  aol://aol_screen_name

In other words, the Zimbra host is out, and we're doing transparent
conversion (since IM app internally still uses the brain-dead XMPP
notation).

Also added a nicer UI so you can select the service type from a
drop-down and enter just the screen name in an input field.

http://bugzilla.zimbra.com/show_bug.cgi?id=14156