bug: 18668, 21059
It turns out that putting a mailbox into Maintenance mode invalidates the Mailbox object -- so the nightly backups (which use Maintenance mode) were causing the Mailbox to be invalidated underneath IM, even though it was holding onto a pointer to the Mailbox.

Re-worked the IMPersona--Mailbox relationship so that IM doesn't rely on it's own long-lived Mailbox pointers: all IM access now goes through the MailboxManager.

http://bugzilla.zimbra.com/show_bug.cgi?id=18668
http://bugzilla.zimbra.com/show_bug.cgi?id=21059