bug: 16525
Before doing a forward, preemptively call MimeMessage.saveChanges().  If this fails, create a copy of the message that doesn't recursively call updateHeaders() upon saveChanges() and call Transport.send() on that instead.  This uses double the memory on malformed messages, but it should avoid having a deep-buried misparse throw off the whole forward.

http://bugzilla.zimbra.com/show_bug.cgi?id=16525