bug: 19222
First pass at new MIME parser.  Unlike JavaMail, it does a decent job of dealing with malformed messages.  The major difference in class layout is that MimeMessage, MimeMultipart, and MimeBodyPart all inherit from MimePart.  This makes a *lot* of things much saner.  You *do* have to wrap your head around the idea that a MimeMessage is just a zero-overhead container for another part, but since MimeMessage.getHeader() Does The Right Thing you just need to remember that the payload of a message is MimeMessage.getBodyPart().getContent(), not MimeMessage.getContent().

Seems to parse things properly, and performance is decent though not quite as good as with a Boyer-Moore parser.  Keeps a pointer to the base File/byte[] and just calculates offsets during the parse; gets content from there when necessary.  (When you have an InputStream, you can either just parse the structure or you can have it read into a byte[] at which point the content is also available.)  Handles base64 and QP encoding/decoding, which is performed lazily when a part's encoding is changed.

Still needs real RFC 2047 header support, plus serialization/deserialization of message structure for caching in the DB.

ContentType and ContentDisposition have been broken out from MimeCompoundHeader.  There are 2 copies of each of these classes for now.  A subsequent checkin will remove the ones in com.zimbra.cs.mime.

http://bugzilla.zimbra.com/show_bug.cgi?id=19222