bug: 19221 - Do not hold whole message in memory
MimeMessages that reference a blob on disk now use our own SharedInputStream
implementation called BlobInputStream.  This implementation is similar to
JavaMail's SharedFileInputStream.  The main difference is that each
substream created by the newStream() method opens a separate filehandle.

This approach is less efficient than JavaMail's approach, but it is more
stable.  It allows us to use a cached MimeMessage and create new streams
to the content, even though the original stream may have been closed.  It
also allows the data to be accessed concurrently by multiple threads.

* FileBlobStore.getContent() now returns a SharedByteArrayInputStream or
BlobInputStream, determined by the file size.
* Renamed zimbraLmtpDiskStreamingThreshold to zimbraStoreDiskStreamingThreshold,
since the attribute now used outside LMTP.
* Added a unit test for BlobInputStream to the test suite.
* Fixed a typo in ZimbraLmtpBackend where we would stream to disk by.
* Confirmed that the new code works with JavaMail 1.4ea.

default if the threshold could not be determined.

http://bugzilla.zimbra.com/show_bug.cgi?id=19221