bug: 20616
Reworked how multi-mailbox does sync.  That way a user can switch to one account and click "Get Mail" to just sync that account.  If the account is local account, then "Get Mail" syncs all POP/IMAP datasources.  The scheduled sync still runs on clock and sequentially going through all datasources and all accounts to sync one-by-one.  But if during the scheduled sync a user requests to sync an account by clicking "Get Mail", we run that in parallel as long as the scheduled sync is not currently syncing that particular mailbox.  If the scheduler reaches a mailbox that happens to be in a manual sync cycle, the scheduled sync skips that mailbox.

Also fixed data source sync interval.  Now the default is 5 minutes instead of 60 secondds.  That's because IMAP for the time being is expensive.

http://bugzilla.zimbra.com/show_bug.cgi?id=20616