Bug: 14156
Display presence information in mail msg. view.  This is done in the
Zimlet, since it displays the contacts in the message headers.  We hold
a cache of the last 50 icons so we can update them when buddies change
presence (I couldn't find another way to keep track of what's on the
screen...)

http://bugzilla.zimbra.com/show_bug.cgi?id=14156