bug: 21532
Fix queries like  "in:junk is:local"

Instead of expanding 'is:local' and 'is:remote' when we generate the parse tree, we instead need to expand them to Pseudo-Ops...then we do the trash/spam evaluation on the parse tree, and then finally we can do the expansion.  This is necessary because doing the actual expansion creates an implicit folder-target, which interferes with the spam/trash evaluation.

http://bugzilla.zimbra.com/show_bug.cgi?id=21532