Bug: 20449, 21210
Use "brute force" approach to update an item in the buddy list (remove
from list, add back).  This solves a few problems such as updating the
groups for a buddy or placing a buddy in the right position according
to the current order.

Allow moving buddies between groups with d'n'd.

http://bugzilla.zimbra.com/show_bug.cgi?id=20449
http://bugzilla.zimbra.com/show_bug.cgi?id=21210